Refactor: rename "network" directory
