+++ drivers/device/device.go
@@ -10,7 +10,7 @@ import (
 	"github.com/timoth-y/chainmetric-sensorsys/drivers/periphery"
 	"github.com/timoth-y/chainmetric-sensorsys/drivers/sensor"
 	"github.com/timoth-y/chainmetric-sensorsys/engine"
-	"github.com/timoth-y/chainmetric-sensorsys/gateway/blockchain"
+	"github.com/timoth-y/chainmetric-sensorsys/network/blockchain"
 	"github.com/timoth-y/chainmetric-sensorsys/model"
 )
