+++ network/blockchain/assets.go
